requirements.txt
cassiopeia==5.0.3
matplotlib==3.7.1
pandas==2.0.1
requests==2.28.2
riotwatcher==3.2.4
seaborn==0.12.2
python-decouple==3.8